requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1
scipy>=1.11.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# API + Dashboard
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
streamlit>=1.28.0

# Logging & utilities
tensorboard>=2.14.0
tqdm>=4.66.0
pydantic>=2.4.0

# Optional: ONNX export
# onnx>=1.15.0
# onnxruntime>=1.16.0
